Renamed: component PickerPriceRanges.vue
