+++ apps/frontend/src/components/page-products/PickerPriceRanges.vue
